Make API work with mobile workers too
