+++ corehq/apps/locations/permissions.py
@@ -3,7 +3,6 @@ from django.http import Http404
 from .models import SQLLocation
 from corehq.apps.domain.decorators import (login_and_domain_required,
                                            domain_admin_required)
-from corehq.util.quickcache import quickcache
 
 
 def locations_access_required(view_fn):
@@ -21,21 +20,6 @@ def is_locations_admin(view_fn):
     return locations_access_required(domain_admin_required(view_fn))
 
 
-@quickcache(['user._id', 'project.name'])
-def editable_locations_ids(user, project):
-    if (user.is_domain_admin(project.name) or
-            not project.location_restriction_for_users):
-        return (SQLLocation.by_domain(project.name)
-                           .values_list('location_id', flat=True))
-
-    user_loc = user.get_location(project.name)
-    if not user_loc:
-        return []
-
-    return list(user_loc.sql_location.get_descendants(include_self=True)
-                                     .values_list('location_id', flat=True))
-
-
 def user_can_edit_location(user, sql_location, project):
     if (user.is_domain_admin(project.name) or
             not project.location_restriction_for_users):
@@ -45,21 +29,6 @@ def user_can_edit_location(user, sql_location, project):
     if user_loc:
         user_loc = user_loc.sql_location
     return user_loc is None or user_loc.is_direct_ancestor_of(sql_location)
-
-
-def viewable_locations_ids(user, project):
-    if (user.is_domain_admin(project.name) or
-            not project.location_restriction_for_users):
-        return (SQLLocation.by_domain(project.name)
-                           .values_list('location_id', flat=True))
-
-    user_loc = user.get_location(project.name)
-    if not user_loc:
-        return []
-
-    return (list(user_loc.sql_location.get_ancestors()
-            .values_list('location_id', flat=True)) +
-            editable_locations_ids(user, project))
 
 
 def user_can_view_location(user, sql_location, project):
